Update link to podcast (#736)
* Update link to podcast

The link to the podcast referred to the podcast homepage, not a specific episode. It took me a whole episode (the latest one) to realize I was listening to the wrong one. 😅 The new link refers to a timestamp in the correct podcast, where verbosity is discussed.

Also corrected a typo.

* Revert update of second podcast link

* update md

---------

Co-authored-by: Maëlle Salmon <maelle.salmon@yahoo.se>